chore: move layout to app root
