+++ apps/web/src/app/(app)/layout.tsx
@@ -1,9 +1,9 @@
 import '~/styles/globals.css'
 
 import { Suspense } from 'react'
-import { Footer } from '~/app/(app)/(public)/(layout)/Footer'
-import { NavBar } from '~/app/(app)/(public)/(layout)/nav/NavBar'
-import { NavBarSkeleton } from '~/app/(app)/(public)/(layout)/nav/NavigationClient'
+import { Footer } from '~/app/(app)/(layout)/Footer'
+import { NavBar } from '~/app/(app)/(layout)/nav/NavBar'
+import { NavBarSkeleton } from '~/app/(app)/(layout)/nav/NavigationClient'
 
 const RootLayout = ({ children }: Readonly<{ children: React.ReactNode }>) => {
   return (
@@ -12,7 +12,7 @@ const RootLayout = ({ children }: Readonly<{ children: React.ReactNode }>) => {
         <NavBar />
       </Suspense>
 
-      <main className="from-background via-muted/30 to-secondary/20 flex-1 bg-gradient-to-br">
+      <main className="from-background via-muted/30 to-secondary/20 flex-1 bg-gradient-to-br pt-16">
         <div className="min-h-full">{children}</div>
       </main>
