+++ apps/web/src/app/(auth)/auth/page.tsx
@@ -1,4 +1,4 @@
-import { Footer } from '~/app/(app)/(public)/(layout)/Footer'
+import { Footer } from '~/app/(app)/(layout)/Footer'
 import { WaitlistPage } from '~/app/(auth)/auth/WaitlistPage'
 
 const AuthPage = async () => {
